Refactor the structure of joint_embedding_scvi and merged with main branch
