+++ CHANGELOG.md
@@ -16,7 +16,7 @@ to [Semantic Versioning]. The full commit history is available in the [commit lo
 - Add support for rapids-singlecell, {pr}`3811`.
 - Add {class}`scvi.external.CytoVI` KNN imputation backend option to be cuML, {pr}`3821`.
 - Add {class}`scvi.model.JointEmbeddingSCVI`, a self-supervised SCVI variant using binomial
-    thinning and a cross-correlation objective (CCO) for robust embeddings, {pr}`XXXX`.
+    thinning and a cross-correlation objective (CCO) for robust embeddings, {pr}`3883`.
 
 #### Fixed
