+++ docs/references.bib
@@ -386,3 +386,12 @@
   month = apr,
   publisher={Nature Publishing Group}
 }
+
+@article{Svensson26,
+  title={Improving SCVI for low-count cells through self-supervised augmentation},
+  author={Valentine Svensson},
+  journal={bioRxiv},
+  year={2026},
+  publisher={Cold Spring Harbor Laboratory},
+  doi = {10.64898/2026.02.11.705441}
+}
