+++ docs/user_guide/models/index.md
@@ -12,6 +12,7 @@ decipher
 destvi
 diagvi
 gimvi
+jointembeddingscvi
 linearscvi
 methylanvi
 methylvi
