+++ docs/user_guide/models/jointembeddingscvi.md
@@ -1,0 +1,20 @@
+# Joint-Embedding SCVI
+
+:::{note}
+This page is under construction.
+:::
+
+**JointEmbeddingSCVI** {cite:p}`Svensson26` (Python class {class}`~scvi.external.JointEmbeddingSCVI`) is a ...
+
+The advantages of JointEmbeddingSCVI are:
+
+-   ...
+
+The limitations of JointEmbeddingSCVI include:
+
+-   ...
+
+```{topic} Tutorials:
+
+-   {doc}`/tutorials/notebooks/scrna/JointEmbeddingSCVI_tutorial`
+```
