+++ src/scvi/external/__init__.py
@@ -4,6 +4,7 @@ from .cytovi import CYTOVI
 from .decipher import Decipher
 from .diagvi import DIAGVI
 from .gimvi import GIMVI
+from .joint_embedding_scvi import JointEmbeddingSCVI
 from .methylvi import METHYLANVI, METHYLVI
 from .mrvi import MRVI
 from .poissonvi import POISSONVI
@@ -22,6 +23,7 @@ __all__ = [
     "SCAR",
     "SOLO",
     "GIMVI",
+    "JointEmbeddingSCVI",
     "Decipher",
     "RNAStereoscope",
     "SpatialStereoscope",
