+++ src/scvi/external/joint_embedding_scvi/__init__.py
@@ -1,0 +1,4 @@
+from ._model import JointEmbeddingSCVI
+from ._module import JointEmbeddingVAE
+
+__all__ = ["JointEmbeddingSCVI", "JointEmbeddingVAE"]
