+++ src/scvi/module/__init__.py
@@ -1,12 +1,12 @@
 import warnings
 
 from scvi import settings
+from scvi.external.joint_embedding_scvi._joint_embedding_vae import JointEmbeddingVAE
 from scvi.utils import error_on_missing_dependencies
 
 from ._amortizedlda import AmortizedLDAPyroModule
 from ._autozivae import AutoZIVAE
 from ._classifier import Classifier
-from ._joint_embedding_vae import JointEmbeddingVAE
 from ._mrdeconv import MRDeconv
 from ._multivae import MULTIVAE
 from ._peakvae import PEAKVAE
